+++ includes/class-tmsm-admin-cleanup.php
@@ -203,9 +203,11 @@ class Tmsm_Admin_Cleanup {
 		$this->loader->add_filter( 'content_block_post_type', $plugin_admin, 'content_block_post_type_public', 10 );
 
 		// Gravity Forms
+
 		$this->loader->add_filter( 'gform_enable_field_label_visibility_settings', $plugin_admin, 'gravityforms_label_visibility', 10 ); // Label Visibility
 
 		// WooCommerce
+		$this->loader->add_filter( 'woocommerce_helper_suppress_admin_notices', $plugin_admin, 'woocommerce_helper_suppress_admin_notices', 10 ); // Remove notices
 		$this->loader->add_filter( 'woocommerce_enable_admin_help_tab', $plugin_admin, 'woocommerce_enable_admin_help_tab' );
 		$this->loader->add_action( 'woocommerce_admin_process_product_object', $plugin_admin, 'wprocket_empty_cache_on_save_product' );
 		$this->loader->add_filter( 'woocommerce_admin_order_date_format', $plugin_admin, 'woocommerce_admin_order_date_format' );